requirements.txt
# Core Dependencies
langgraph==0.2.16
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0

# LLM & AI
transformers==4.45.0
torch==2.4.0
accelerate==0.34.0
bitsandbytes==0.44.0
peft==0.12.0
trl==0.10.0  # For DPO training
datasets==2.21.0
sentencepiece==0.2.0

# API & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9

# Gmail & Google APIs
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0

# Database & Storage
chromadb==0.5.5
sqlalchemy==2.0.35
alembic==1.13.3

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
aiohttp==3.10.5
tenacity==9.0.0
redis==5.1.1
parsedatetime==2.6
python-dateutil==2.9.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.21.0

# Testing & Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2
pre-commit==3.8.0

# Visualization & Analysis
pandas==2.2.3
matplotlib==3.9.2
plotly==5.24.0

